Multiple checkbox supports schema.items.enum
The previous way of handling checkbox multiple values was with an "enum" list in the schema. However, when the type is "array" (as it probably should be), the "enum" should be a child of the "items", rather than parent. Otherwise, JSON Schema validation should technically fail.

Doing things this way allows a multi-select list to quickly change from a "select" type to a "checkbox" type by changing one field in options.