Adding order option to ObjectField.
As discussed in https://github.com/gitana/alpaca/issues/249 there is no
guarantee that fields will be rendered in the order they are delcared in
the schema. By adding an `order` option, users can optionally choose
whether or not to enforce a certain order without having to create a
layout.