Improve enum validation error.
When a required field has an enum of possible values, the
values are reported back to the user if nothing is selected.
However, when the field also has an optionLabels array the
user only sees the labels in the UI. As a result, the error
message shows values that are not presented to the user.

This change uses the optionLabels in the error message if
they exist.