Fixed required select field with visible "none" option
A required select field with a visible none option, selected the first real option as start value in stead of the none option.

When the select is configured like this, we want the select to start with the "watermark/none" option selected and force the user to select one of the real options. This used to be the behaviour of the select field.