Fixing multiselect display bug for selects not using multiselect
If you had multiselect library loaded, but were not using it for a sepecific select, then you disable/enable it, it would switch from normal select to multiselect.

I mimiced the logic used elsewhere to decide whether or not to render multiselect, which is whether or not the 'multiple' option was set.

see jsfiddle https://jsfiddle.net/anneb574/rhce3oxf/